+++ iothub_client/tests/iothubclient_mqtt_mod_dm_e2e/main.c
@@ -7,5 +7,5 @@ int main(void)
 {
     size_t failedTestCount = 0;
     RUN_TEST_SUITE(iothubclient_mqtt_mod_dm_e2e, failedTestCount);
-    return failedTestCount;
+    return (int)failedTestCount;
 }
